==> Final Project/Data wrangling/1 - Combining JSON files.ipynb <==
# The purpose of this notebook is to combine the rightmove json files that were generated from the rightmove script
# The script leverages the Borough.xlsx file to add a borough column to the dataset

import pandas as pd
from pathlib import Path
import numpy as np
import re


#Path to the list of London boroughs that are being modelled
PATH_BOROUGHS = Path(r'..\Rightmove Dumps JSON\Boroughs.xlsx')
#Path to the list of JSON files pulled from Rightmove.co.uk
PATH_JSON = r'..\Rightmove Dumps JSON'
#Name of the concatenated file that will be used for analysis
RIGHTMOVE_FILE = 'right_move_data_eda.csv'
FINAL_DATA_PATH = r'../Data'

#Read in the list of London boroughs as a dataframe and convert it to a list
boroughs = pd.read_excel(PATH_BOROUGHS)
boroughs = boroughs.values.tolist()


#Append .json to the end of each borough name
new_borough = []
for borough in boroughs:
    new_borough.append(borough[0] + '.json')

#concatenate the list of list of .JSON files into a single dataframe
dfs = []
for file in new_borough:
    data = pd.read_json(Path(PATH_JSON)/file)
    # knowing the bourough of each property may be useful
    data['bourough'] = file
    dfs.append(data)
df = pd.concat(dfs, ignore_index=True)
df.bouroughv(Path(FINAL_DATA_PATH)/RIGHTMOVE_FILE)
