==> Final Project/Data wrangling/2 - Inital EDA.ipynb <==
import pandas as pd

ORIGINAL_FILE = r'../Data/rm_working_files/right_move_data_eda.csv'
UPDATED_FILE = r'../Data/rm_working_files/right_move_data_eda_update.csv'

feature_columns = ['price', 'bedrooms', 'address', 'latitude', 'longitude', 'propertyType','summary', 'bourough']
rm = pd.read_csv(ORIGINAL_FILE, usecols=feature_columns)

rm.info()

rm.shape

# - Looks like there are a number of duplicate property types in the list which is being caused by capital letters - 
# - Remove all caps 
rm.propertyType.value_counts()

rm = rm.apply(lambda x: x.str.lower()if x.dtype =='object' else x)

rm.info()

#There are 473 property listings which do not have details re the number of bedrooms
rm.isnull().sum()

# update studio flat and Ground floor studio flat  = studio apartment as they are the same 
# 321 of the 473 rows where bedroom = NaN are studio apartments so I can update these records to have 0 bedrooms
rm.loc[rm['propertyType'].isin(['studio flat', 'ground floor studio flat']), 'propertyType'] = 'studio apartment'
rm.loc[(rm['bedrooms'].isnull()) & (rm['propertyType'] == 'studio apartment'), 'bedrooms'] = 0

# I want to see the remaining bedrooms = nan and their corresponding property types

rm[rm.bedrooms.isnull()].groupby(rm['propertyType']).count()

# remove the following property types from the dataset: Land, Parking, Plot, Block of apartments, hotel rooms
rm.drop(rm.loc[rm['propertyType'].isin(['land', 'parking', 'plot','block of apartments','garages', 'hotel room', 'house boat', 'park home'])].index, inplace=True)


# the remaining property types that do have bedrooms = NaN are as follows:

rm[rm.bedrooms.isnull()].groupby(rm['propertyType']).count()

rm.info()

# Remove the 61 rows where bedroom = na
# because the bedroom column is the only one with na, we can simply use the dropna function

rm.dropna(inplace=True)

rm.isnull().sum()

#See of garden is mentioned in the summary 
rm['outdoor space'] = rm['summary'].str.contains('balcony|garden|terrace|yard|acre',regex=True)

#remove .json from the end of each bourough 
# rm.bourough.str.replace('.json', '')
# df['prod_type'] = df['prod_type'].replace('respon' ,'responsvie')
rm.bourough.replace('.json', '', inplace=True, regex=True)

rm.to_csv(r'../Data wrangling/ids.csv')

# - Looks like there are a number of duplicate property types in the list which is being caused by capital letters - 
# - Remove all caps 
rm.propertyType.value_counts()

#There are too many property types in the list so I have decided to combine the types that are
#similar / the same:
    # apartments, flats, serviced apartment, ground floor flat
    # terraced house, semi-detacthed house, cluster house 
    # detacthed house, house, town house, link detached house
    # duplex,  ground maisonette, triplex 
    # detached bungalow,semi-detached bungalow, 
    # manor house equestrian facility chalet lodge
    
rm.loc[rm.propertyType.isin(['flat','serviced apartment','ground floor flat']), 'propertyType'] = 'apartment' 
rm.loc[rm.propertyType.isin(['terraced house','semi-detacthed house', 'cluster house']), 'propertyType'] = 'terraced house'  
rm.loc[rm.propertyType.isin(['detacthed house','town house', 'house', 'link detached house']), 'propertyType'] = 'detached house'  
rm.loc[rm.propertyType.isin(['duplex','ground maisonette', 'triplex']), 'propertyType'] = 'maisonette'  
rm.loc[rm.propertyType.isin(['detached bungalow','semi-detached bungalow','terraced bungalow']), 'propertyType'] = 'bungalow'  
rm.loc[rm.propertyType.isin(['manor house','equestrian facility', 'chalet', 'lodge']), 'propertyType'] = 'country house'  

rm.propertyType.value_counts()

# property type = 'property' appears to be a mix of houses and apartments
rm[rm['propertyType'] == 'property']

mask = rm['propertyType'] == 'property'
rm.drop(rm[mask].index, inplace=True)

rm.propertyType.value_counts()

#pull the postcodes from the address column, they appear to be the last part of the address string, separated by a comma
rm['postcode'] = rm['address'].str.rsplit(',').str[-1]

#reviewing the postcodes column coloser, it appears that not all postcodes are provided for every address so,
#drop the postcodes column - I will need to find another way of finding these postcodes
#rm.to_csv(r'..\Data Dumps JSON\Examine.csv') 

rm.drop('postcode', axis=1, inplace=True)

rm.head(7)

rm.to_csv(UPDATED_FILE)
