==> Final Project/Data wrangling/3 - Adding postcode .ipynb <==
# feature 1 = I will require the postcodes for each property in the dataset. Initial analysis of the address column has 
# uncovered that the postcodes cannot be used as they are partially populated. 

# I can leverage the long and lat within the rm dataset and the https://postcodes.io/ API to
#return the postcode for each row in the dataset, 

import requests
import pandas as pd

ORIGINAL_FILE = r'../Data/rm_working_files/right_move_data_eda_update.csv'
UPDATED_FILE = r'../Data/rm_working_files/right_move_data_eda_postcodes.csv'

df = pd.read_csv(ORIGINAL_FILE )
postcode = []
for i in range(0, len(df) + 1):

    LONG = df['longitude'][i]
    LAT = df['latitude'][i]
    API = f'https://api.postcodes.io/postcodes?lon={LONG}&lat={LAT}'

    r = requests.get(API)
    # x = r.json() - need to understand why the script is thowing an error
    
    if r.json()['result'] != None:
        postcode.append([r.json()['result'][0]['postcode']])
    else:
        postcode.append('Not available')
postcode = pd.DataFrame(postcode, columns='postcode')
df1 = pd.DataFrame({'postcode': postcode})
df1.to_csv(UPDATED_FILE)
df_postcode = pd.concat([df, df1], axis=1)
df_postcode.to_csv(UPDATED_FILE)

postcodes = pd.read_csv(r'../Data/postcodes.csv')

rm = pd.read_csv(ORIGINAL_FILE)

rm = pd.concat([rm,postcodes], axis=1)
rm.head(19)

rm['postcode'] = rm['postcode'].map(lambda x: x.lstrip("['").rstrip("']"))

# there were some instance where postcode was not returned from the API, in these cases attempt
#to pull the postcode from the address column
rm.loc[rm.postcode == 'Not available', 'postcode'] = rm['address'].str.rsplit(',').str[-1]

#Only require the first part of the postcode
rm.postcode.str.upper()
rm['postcode first section'] = rm['postcode'].str.split().str[0]

rm.head()

rm.to_csv(UPDATED_FILE)
