==> Final Project/Data wrangling/4 - Adding features.ipynb <==
#Adding the following features to the data set:
# from tfl data, add the nearest station and total number of stations within a 2km radius
# from schools data, add the total number of schools per borough and the percentage of students who attend private school per borough
# add median income per person in each borough
# add the population density for each borough

import pandas as pd


ORIGINAL_FILE = r'../Data/rm_working_files/right_move_data_eda_postcodes.csv'

#Features to be added: 
TFL_FILE = r'../Data/tfl/tfl_data_final.csv'
POPULATION_DENSITY = r'../Data/population_densiity/population_density.csv'
MEDIAN_EARNING_PER_BOUROUHG = r'../Data/earnings/earnings-residence-borough.csv'
SCHOOLS = r'../Data/schools/schools_per_borough.csv'
UPDATED_FILE = r'../Data/rm_working_files/right_move_data_with_features.csv'

rm = pd.read_csv(ORIGINAL_FILE)

tfl = pd.read_csv(TFL_FILE)

pop = pd.read_csv(POPULATION_DENSITY)

earnings = pd.read_csv(MEDIAN_EARNING_PER_BOUROUHG)

schools = pd.read_csv(SCHOOLS)

rm_tube = pd.concat([rm, tfl], axis = 1)

rm_tube.info()

pop.bourough = pop.bourough.str.lower()

pop.head()

# Adding population density to the dataframe
rm_pop = pd.merge(left=rm_tube, right=pop, on=['bourough'], how='left')

rm_pop.isnull().sum()

earnings.head()

rm_wage = pd.merge(left=rm_pop, right=earnings, on=['bourough'],how='left')

rm_wage.isnull().sum()

schools.bourough = schools.bourough.str.lower()

rm_wage = pd.merge(left=rm_wage, right=schools, on='bourough', how='left')

rm_wage.head()

rm_wage.to_csv(UPDATED_FILE)

rm = pd.read_csv(UPDATED_FILE)

rm.info()

rm.drop_duplicates()

rm.to_csv(UPDATED_FILE)
