==> Final Project/Data wrangling/5 - Final tidy up.ipynb <==
import warnings
warnings.filterwarnings("ignore")
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.metrics import r2_score

DATA_FILE = r'../Data/rm_working_files/rm_working_files/right_move_data_final1.csv'
DATA_READY = r'../Data/rm_working_files/right_move_data_final2.csv'

df = pd.read_csv(DATA_FILE, thousands=' ')

df['median wage'].replace(',' , '')

#df["a"] = pd.to_numeric(df["a"])
df['nearest stattion'] = df['nearest stattion'].replace('Nothing within a 2km radius', 0)
df['nearest stattion'].astype(float)

df['median wage'].value_counts()

df['population'].head()

df['population'].astype(int)

df.info()

df.drop_duplicates(subset=['postcode','price'], inplace=True)

df.to_csv(DATA_READY)
